tweaks: posts pagination re-write
