+++ src/hooks/usePagination.ts
@@ -1,4 +1,4 @@
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 import useContractState from '@/hooks/useContractState';
 import { Ordering, Pagination } from '@/types';
 import { stringToNum } from '@/utils/number';
@@ -8,16 +8,20 @@ export default function usePagination<T>(
   contract: ChainContract | undefined,
   paginationMessage: string,
   recordPerPage: number,
+  descending = false,
+  nonce = 0,
 ) {
   const [pageIndex, setPageIndex] = useState(1);
-  const { state: page } = useContractState<Pagination<T>>(
-    contract,
-    paginationMessage,
-    useMemo(() => [(pageIndex - 1) * recordPerPage, recordPerPage, { Descending: ['12'] }], [pageIndex, recordPerPage]),
-  );
 
-  console.log([(pageIndex - 1) * recordPerPage, recordPerPage, { Descending: ['12'] }]);
-  console.log(page);
+  // Nonce only for descending
+  let args: unknown[];
+  if (descending) {
+    args = [nonce - (pageIndex - 1) * recordPerPage, recordPerPage, Ordering.Descending];
+  } else {
+    args = [(pageIndex - 1) * recordPerPage, recordPerPage];
+  }
+
+  const { state: page } = useContractState<Pagination<T>>(contract, paginationMessage, args);
 
   const { items, total } = page || {};
   const numberOfPage = stringToNum(total) ? Math.ceil(parseInt(total!) / recordPerPage) : 1;
