+++ src/pages/plugins/Posts/index.tsx
@@ -1,71 +1,8 @@
-import { Box, Flex, Tag, Text } from '@chakra-ui/react';
-import { useEffect, useState } from 'react';
-import { useWindowScroll } from 'react-use';
-import usePagination from '@/hooks/usePagination';
-import PostCard from '@/pages/plugins/Posts/PostCard';
-import PostsProvider, { usePostsContext } from '@/pages/plugins/Posts/PostsProvider';
-import NewPostButton from '@/pages/plugins/Posts/action/NewPostButton';
+import { Box, Text } from '@chakra-ui/react';
+import PostsProvider from '@/pages/plugins/Posts/PostsProvider';
+import PostsView from '@/pages/plugins/Posts/PostsView';
 import { useSpaceContext } from '@/providers/SpaceProvider';
-import { MemberStatus, Ordering, PostRecord } from '@/types';
 import { PLUGIN_POSTS } from '@/utils/plugins';
-
-const RECORD_PER_PAGE = 5;
-
-function PostsContent() {
-  const { memberStatus } = useSpaceContext();
-  const { contract } = usePostsContext();
-  const [onLoad, setOnLoad] = useState(true);
-  const [storage, setStorage] = useState<PostRecord[]>([]);
-  const {
-    items,
-    pageIndex,
-    setPageIndex,
-    hasNextPage,
-    total: numOfPost,
-  } = usePagination<PostRecord>(contract, 'listPosts', RECORD_PER_PAGE);
-  const { y } = useWindowScroll();
-
-  console.log(items);
-
-  useEffect(() => {
-    if (items && onLoad) {
-      setOnLoad(false);
-      setStorage((prevState) => [...prevState, ...items]);
-    }
-  }, [items]);
-
-  useEffect(() => {
-    if (onLoad || !hasNextPage) return;
-
-    // When the current view bottom -> the bottom of web <= 50 pixel
-    if (document.body.offsetHeight - y - innerHeight <= 50) {
-      setOnLoad(true);
-      setPageIndex(pageIndex + 1);
-    }
-  }, [onLoad, y]);
-
-  return (
-    <Box>
-      <Flex justify='space-between' align='center' mb={4} gap={2}>
-        <Flex gap={2} align='center'>
-          <Text fontSize='xl' fontWeight='semibold'>
-            Posts
-          </Text>
-          <Box>
-            <Tag>{numOfPost}</Tag>
-          </Box>
-        </Flex>
-        <Flex gap={2}>{memberStatus === MemberStatus.Active && <NewPostButton />}</Flex>
-      </Flex>
-
-      <Box>
-        {storage?.map((postRecord) => (
-          <PostCard key={postRecord.postId} postRecord={postRecord} />
-        ))}
-      </Box>
-    </Box>
-  );
-}
 
 export default function Posts() {
   const { plugins } = useSpaceContext();
@@ -84,7 +21,7 @@ export default function Posts() {
 
   return (
     <PostsProvider info={postPlugin}>
-      <PostsContent />
+      <PostsView />
     </PostsProvider>
   );
 }
